refactor(app): getLabwareDefinitionUri to xforms
This takes a full protocol, which is why it's here. Could see an
argument for it being in local-resources somewhere too.
