+++ app/src/molecules/PythonLabwareOffsetSnippet/createSnippet.ts
@@ -1,6 +1,6 @@
 import isEqual from 'lodash/isEqual'
 import { getLoadedLabwareDefinitionsByUri } from '@opentrons/shared-data'
-import { getLabwareDefinitionUri } from '/app/organisms/Devices/ProtocolRun/utils/getLabwareDefinitionUri'
+import { getLabwareDefinitionUri } from '/app/transformations/protocols'
 import { getLabwareOffsetLocation } from '/app/organisms/Devices/ProtocolRun/utils/getLabwareOffsetLocation'
 import type { LabwareOffset } from '@opentrons/api-client'
 import type {
