+++ app/src/organisms/Devices/ProtocolRun/useLabwareOffsetForLabware.ts
@@ -1,7 +1,7 @@
 import { getLoadedLabwareDefinitionsByUri } from '@opentrons/shared-data'
 
 import { getCurrentOffsetForLabwareInLocation } from './utils/getCurrentOffsetForLabwareInLocation'
-import { getLabwareDefinitionUri } from './utils/getLabwareDefinitionUri'
+import { getLabwareDefinitionUri } from '/app/transformations/protocols'
 import { getLabwareOffsetLocation } from './utils/getLabwareOffsetLocation'
 import { useNotifyRunQuery } from '/app/resources/runs'
